+++ app/components/Authentication/Login.tsx
@@ -2,10 +2,7 @@ import { ChangeEvent, useActionState, useEffect, useState } from "react";
 import { Input } from "@/components/ui/input";
 import { Label } from "@/components/ui/label";
 import { CardContent } from "@/components/ui/card";
-import {
-  useAuthenticationAction,
-  useToggle,
-} from "@/app/context/ContextProvider";
+import { useToggle } from "@/app/context/PopupContext";
 import { login } from "./actions/login";
 import { FormError, PropsWithReactNode } from "./type";
 
@@ -13,15 +10,13 @@ export default function Login({ header, button, onClose }: PropsWithReactNode) {
   const [userInput, handleInput] = useInput(initialData);
   const [formState, formAction] = useActionState(login, initialFormError);
   const toggle = useToggle();
-  const loginUser = useAuthenticationAction();
 
   useEffect(() => {
     if (formState.state === "success") {
       toggle();
       onClose();
-      loginUser();
     }
-  }, [toggle, onClose, loginUser, formState.state]);
+  }, [toggle, onClose, formState.state]);
 
   return (
     <>
