+++ app/components/Authentication/Signup.tsx
@@ -2,7 +2,7 @@ import { ChangeEvent, useActionState, useEffect, useState } from "react";
 import { Input } from "@/components/ui/input";
 import { Label } from "@/components/ui/label";
 import { CardContent } from "@/components/ui/card";
-import { useToggle } from "@/app/context/ContextProvider";
+import { useToggle } from "@/app/context/PopupContext";
 import { signup } from "./actions/signup";
 import { FormError, PropsWithReactNode } from "./type";
