+++ app/layout.tsx
@@ -1,8 +1,7 @@
 import type { Metadata } from "next";
 import { Geist, Geist_Mono } from "next/font/google";
 import "./globals.css";
-import { cookies } from "next/headers";
-import ContextProvider from "./context/ContextProvider";
+import PopupProvider from "./context/PopupContext";
 import Authentication from "./components/Authentication/Authentication";
 import Header from "./(home)/components/Header";
 
@@ -26,19 +25,16 @@ export default async function RootLayout({
 }: Readonly<{
   children: React.ReactNode;
 }>) {
-  const session = (await cookies()).get("session");
-  const isLogin = session !== undefined;
-
   return (
     <html lang="ko">
       <body
         className={`${geistSans.variable} ${geistMono.variable} antialiased`}
       >
-        <ContextProvider initialUserState={isLogin}>
+        <PopupProvider>
           <Authentication />
           <Header />
           {children}
-        </ContextProvider>
+        </PopupProvider>
       </body>
     </html>
   );
